packages/processor/pyproject.toml
[project]
name = "langgate-processor"
version = "0.1.0"
description = "Envoy External Processor for LangGate AI Gateway"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Tanantor Ltd." }]
maintainers = [{ name = "Saran Connoly", email = "saran@tanantor.com" }]
license = { text = "MIT" }
dependencies = [
    "langgate-core (>=0.1.0,<0.2.0)",
    "langgate-transform (>=0.1.0,<0.2.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "protobuf (>=6.30.2,<7.0.0)",
]

[project.optional-dependencies]
dev = [
    "grpc-stubs (>=1.53.0.5,<2.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langgate"]
